Code review fixes, refactoring for its modules and update wasm version.
