+++ auth/wasm/Cargo.toml
@@ -25,7 +25,7 @@ overflow-checks = false
 path = ".."
 
 [dependencies.multiversx-sc-wasm-adapter]
-version = "0.46.1"
+version = "0.47.7"
 
 [workspace]
 members = ["."]
