+++ interchain-token-service/src/constants.rs
@@ -213,3 +213,10 @@ impl<M: ManagedTypeApi> AbiEncodeDecode<M> for DeployTokenManagerPayload<M> {
         }
     }
 }
+
+pub struct TransferAndGasTokens<M: ManagedTypeApi> {
+    pub transfer_token: EgldOrEsdtTokenIdentifier<M>,
+    pub transfer_amount: BigUint<M>,
+    pub gas_token: EgldOrEsdtTokenIdentifier<M>,
+    pub gas_amount: BigUint<M>,
+}
