+++ interchain-token-service/src/remote.rs
@@ -4,10 +4,11 @@ use token_manager::constants::TokenManagerType;
 use crate::abi::AbiEncodeDecode;
 use crate::constants::{
     DeployInterchainTokenPayload, DeployTokenManagerPayload, InterchainTransferPayload, Metadata,
-    MetadataVersion, TokenId, LATEST_METADATA_VERSION, MESSAGE_TYPE_DEPLOY_INTERCHAIN_TOKEN,
-    MESSAGE_TYPE_DEPLOY_TOKEN_MANAGER, MESSAGE_TYPE_INTERCHAIN_TRANSFER,
+    MetadataVersion, TokenId, TransferAndGasTokens, LATEST_METADATA_VERSION,
+    MESSAGE_TYPE_DEPLOY_INTERCHAIN_TOKEN, MESSAGE_TYPE_DEPLOY_TOKEN_MANAGER,
+    MESSAGE_TYPE_INTERCHAIN_TRANSFER,
 };
-use crate::{address_tracker, events, express_executor_tracker, proxy};
+use crate::{address_tracker, events, express_executor_tracker, proxy_cgp, proxy_its};
 
 multiversx_sc::imports!();
 
@@ -16,7 +17,8 @@ pub trait RemoteModule:
     express_executor_tracker::ExpressExecutorTracker
     + multiversx_sc_modules::pause::PauseModule
     + events::EventsModule
-    + proxy::ProxyModule
+    + proxy_cgp::ProxyCgpModule
+    + proxy_its::ProxyItsModule
     + address_tracker::AddressTracker
 {
     fn deploy_remote_token_manager(
@@ -103,11 +105,9 @@ pub trait RemoteModule:
         source_address: ManagedAddress,
         destination_chain: ManagedBuffer,
         destination_address: ManagedBuffer,
-        amount: BigUint,
+        transfer_and_gas_tokens: TransferAndGasTokens<Self::Api>,
         metadata_version: MetadataVersion,
         data: ManagedBuffer,
-        gas_token: EgldOrEsdtTokenIdentifier,
-        gas_value: BigUint,
     ) {
         let data_hash = if data.is_empty() {
             ManagedByteArray::from(&[0; KECCAK256_RESULT_LEN])
@@ -120,7 +120,7 @@ pub trait RemoteModule:
             token_id: token_id.clone(),
             source_address: source_address.as_managed_buffer().clone(),
             destination_address: destination_address.clone(),
-            amount: amount.clone(),
+            amount: transfer_and_gas_tokens.transfer_amount.clone(),
             data,
         };
 
@@ -130,8 +130,8 @@ pub trait RemoteModule:
             &destination_chain,
             &payload,
             metadata_version,
-            gas_token,
-            gas_value,
+            transfer_and_gas_tokens.gas_token,
+            transfer_and_gas_tokens.gas_amount,
         );
 
         self.emit_interchain_transfer_event(
@@ -139,7 +139,7 @@ pub trait RemoteModule:
             source_address,
             destination_chain,
             destination_address,
-            amount,
+            transfer_and_gas_tokens.transfer_amount,
             data_hash,
         );
     }
