+++ interchain-token-service/wasm/src/lib.rs
@@ -22,17 +22,17 @@ multiversx_sc_wasm_adapter::endpoints! {
         init => init
         upgrade => upgrade
         setInterchainTokenFactory => set_interchain_token_factory
+        setFlowLimits => set_flow_limits
+        execute => execute
+        contractCallValue => contract_call_value
+        chainNameHash => chain_name_hash
         deployTokenManager => deploy_token_manager
         deployInterchainToken => deploy_interchain_token
         expressExecute => express_execute_endpoint
         interchainTransfer => interchain_transfer
         callContractWithInterchainToken => call_contract_with_interchain_token
-        setFlowLimits => set_flow_limits
-        execute => execute
         interchainTokenId => interchain_token_id
-        contractCallValue => contract_call_value
         interchainTokenFactory => interchain_token_factory
-        chainNameHash => chain_name_hash
         transferOperatorship => transfer_operatorship
         proposeOperatorship => propose_operatorship
         acceptOperatorship => accept_operatorship
@@ -44,14 +44,14 @@ multiversx_sc_wasm_adapter::endpoints! {
         chainName => chain_name
         trustedAddress => trusted_address
         trustedAddressHash => trusted_address_hash
+        gateway => gateway
+        gasService => gas_service
         flowLimit => flow_limit
         flowOutAmount => flow_out_amount
         flowInAmount => flow_in_amount
         validTokenManagerAddress => valid_token_manager_address
         validTokenIdentifier => valid_token_identifier
         invalidTokenManagerAddress => invalid_token_manager_address
-        gateway => gateway
-        gasService => gas_service
         tokenManagerAddress => token_manager_address
         tokenManagerImplementation => token_manager
         pause => pause_endpoint
