+++ ping-pong-interchain/Cargo.toml
@@ -9,7 +9,7 @@ publish = false
 path = "src/ping_pong.rs"
 
 [dependencies.multiversx-sc]
-version = "0.46.1"
+version = "0.47.7"
 
 [dev-dependencies.multiversx-sc-scenario]
-version = "0.46.1"
+version = "0.47.7"
