+++ ping-pong-interchain/meta/Cargo.toml
@@ -11,4 +11,4 @@ publish = false
 path = ".."
 
 [dependencies.multiversx-sc-meta]
-version = "0.46.1"
+version = "0.47.7"
