+++ token-manager/Cargo.toml
@@ -9,10 +9,10 @@ publish = false
 path = "src/lib.rs"
 
 [dependencies.multiversx-sc]
-version = "0.46.1"
+version = "0.47.7"
 
 [dependencies.operatable]
 path = "../modules/operatable"
 
 [dev-dependencies.multiversx-sc-scenario]
-version = "0.46.1"
+version = "0.47.7"
